+++ nova2d/nova-zero/src/core/NovaCore.cpp
@@ -415,12 +415,12 @@ namespace novazero
 
 		float n2dDegToRad(float deg)
 		{
-			return (deg * 3.141592653589793238463) / 180;
+			return (float)((deg * 3.141592653589793238463) / 180);
 		}
 
 		float n2dRadToDeg(float rad)
 		{
-			return rad * (180.0 / 3.141592653589793238463);
+			return (float)(rad * (180.0 / 3.141592653589793238463));
 		}
 
 	}
